[AS7-5286] Provide BundleContext through framework services
