+++ osgi/service/src/main/java/org/jboss/as/osgi/deployment/BundleResolveProcessor.java
@@ -57,7 +57,7 @@ public class BundleResolveProcessor implements DeploymentUnitProcessor {
 
         DeploymentUnit depUnit = phaseContext.getDeploymentUnit();
         Deployment deployment = depUnit.getAttachment(OSGiConstants.DEPLOYMENT_KEY);
-        XBundle bundle = depUnit.getAttachment(OSGiConstants.INSTALLED_BUNDLE_KEY);
+        XBundle bundle = depUnit.getAttachment(OSGiConstants.BUNDLE_KEY);
         if (bundle == null || !deployment.isAutoStart())
             return;
