+++ pom.xml
@@ -181,11 +181,11 @@
         <version.org.jboss.msc.jboss-msc>1.0.2.GA</version.org.jboss.msc.jboss-msc>
         <version.org.jboss.netty>3.2.6.Final</version.org.jboss.netty>
         <version.org.jboss.osgi.deployment>1.1.1.Final</version.org.jboss.osgi.deployment>
-        <version.org.jboss.osgi.framework>2.0.0.CR15</version.org.jboss.osgi.framework>
+        <version.org.jboss.osgi.framework>2.0.0.CR17</version.org.jboss.osgi.framework>
         <version.org.jboss.osgi.logging>1.0.0.Final</version.org.jboss.osgi.logging>
         <version.org.jboss.osgi.metadata>2.1.1.CR7</version.org.jboss.osgi.metadata>
         <version.org.jboss.osgi.resolver>3.0.0.CR8</version.org.jboss.osgi.resolver>
-        <version.org.jboss.osgi.repository>2.0.0.CR3</version.org.jboss.osgi.repository>
+        <version.org.jboss.osgi.repository>2.0.0</version.org.jboss.osgi.repository>
         <version.org.jboss.osgi.spi>3.1.1.CR6</version.org.jboss.osgi.spi>
         <version.org.jboss.osgi.vfs>1.1.2.CR1</version.org.jboss.osgi.vfs>
         <version.org.jboss.remote-naming>1.0.4.Final</version.org.jboss.remote-naming>
